split driver type spec into multiple files
