+++ packages/driver/test/cypress/support/utils.js
@@ -86,6 +86,23 @@ const getAllFn = (...aliases) => {
   )
 }
 
+export const keyEvents = [
+  'keydown',
+  'keyup',
+  'keypress',
+  'input',
+  'textInput',
+]
+
+export const attachKeyListeners = attachListeners(keyEvents)
+
+// trim new lines at the end of innerText
+// due to changing browser versions implementing
+// this differently
+export const trimInnerText = ($el) => {
+  return _.trimEnd($el.get(0).innerText, '\n')
+}
+
 Cypress.Commands.add('getAll', getAllFn)
 
 const chaiSubset = require('chai-subset')
